Create Startup
Little startup script I wrote for the unix server at my school. This will tell you a time based welcome, give you the list of current users, the number of users, current date and time, and the last login for your user. Because my school is a provincial institute it has a 142 Address for the wireless. Therefore I check if the last address used was a 142 address if yes the I know I logged in from the schools WIFI network if its a 10 address I was on an inside Desktop and anything else I was outside.